requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=2.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py38h0a891b7_1
aiohttp-cors=0.7.0=pypi_0
aiorwlock=1.3.0=pypi_0
aiosignal=1.3.1=pyhd8ed1ab_0
anyio=3.7.1=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.2=py38h06a4308_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
blinker=1.7.0=pyhd8ed1ab_0
brotli-python=1.0.9=py38h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.2=pyhd8ed1ab_0
catkin-pkg=1.0.0=pypi_0
catkin-tools=0.9.4=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=unix_pyh707e725_0
cloudpickle=1.2.2=pypi_0
colorful=0.5.5=pypi_0
comm=0.2.0=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py38hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
cython=3.0.0a10=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
distlib=0.3.7=pypi_0
distro=1.8.0=pypi_0
docutils=0.20.1=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.1=pypi_0
fastapi=0.104.1=pypi_0
fasteners=0.19=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py38h06a4308_0
fonttools=4.44.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py38h5eee18b_0
fsspec=2023.10.0=pypi_0
giflib=5.2.1=h5eee18b_3
glfw=2.6.2=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-api-core=2.13.0=pypi_0
google-auth=2.23.4=pyhca7485f_0
google-auth-oauthlib=1.1.0=pyhd8ed1ab_0
googleapis-common-protos=1.61.0=pypi_0
gpustat=1.1.1=pypi_0
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.48.2=py38he1ff14a_1
gym=0.15.3=pypi_0
gym-notices=0.0.8=pypi_0
h11=0.14.0=pypi_0
higher=0.2.1=pypi_0
httptools=0.6.1=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.32.0=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.26.0=pypi_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.0.34=0
libcurand=10.3.4.52=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lxml=4.9.3=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pyhd8ed1ab_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
msgpack=1.0.7=pypi_0
mujoco-py=2.1.2.14=pypi_0
multidict=6.0.2=py38h5eee18b_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-ml-py=12.535.133=pypi_0
oauthlib=3.2.2=pyhd8ed1ab_0
opencensus=0.11.3=pypi_0
opencensus-context=0.1.3=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
osrf-pycommon=2.0.2=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3=py38h06a4308_0
platformdirs=3.11.0=pypi_0
prometheus-client=0.18.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.25.0=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-spy=0.3.14=pypi_0
pyarrow=14.0.1=pypi_0
pyasn1=0.5.0=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pybullet=3.2.5=pypi_0
pybulletgym=0.1=dev_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.13=pypi_0
pyglet=1.3.2=pypi_0
pygments=2.16.1=pypi_0
pyjwt=2.8.0=pyhd8ed1ab_0
pymavlink=2.4.40=pypi_0
pyopengl=3.1.7=pypi_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
python-dotenv=1.0.0=pypi_0
python-graphviz=0.20.1=pypi_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
ray=2.8.0=pypi_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rospkg=1.5.0=pypi_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.0=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
smart-open=6.4.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.27.0=pypi_0
sympy=1.11.1=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py38h52d8a92_0
tensorboardx=2.4.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.0.2=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu117
torchviz=0.0.2=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.13.0=pypi_0
typing-extensions=4.8.0=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
urllib3=1.26.18=py38h06a4308_0
uvicorn=0.24.0.post1=pypi_0
uvloop=0.19.0=pypi_0
virtualenv=20.21.0=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.9=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.2=h5eee18b_0
yarl=1.7.2=py38h0a891b7_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
